Authorization on bookmarks controller.

+++ app/controllers/bookmarks_controller.rb
@@ -1,13 +1,10 @@
 class BookmarksController < ApplicationController
 
+  before_filter :force_login
+
   def show
     @bookmark = Bookmark.find(params[:id], auth: current_auth)
     raise ActionController::RoutingError.new('Not Found') unless @bookmark
-  end
-
-  def edit
-    redirect_to(bookmarks_path) unless current_auth
-    @bookmark = Bookmark.find(params[:id], auth: current_auth)
   end
 
   def create
@@ -18,20 +15,26 @@ class BookmarksController < ApplicationController
     results.valid? ? redirect_to(:back, notice: t('bookmarks.successfully created')) : render(action: "new")
   end
 
-  def update
-    params[:bookmark][:color] = params[:highlight][:color] if params[:highlight]
-
+  def edit
     @bookmark = Bookmark.find(params[:id], auth: current_auth)
-    @bookmark.auth = current_auth
+    redirect_to(moments_path) and return unless @bookmark.user_id == current_auth.user_id
+  end
 
+  def update
+    @bookmark = Bookmark.find(params[:id], auth: current_auth)
+    redirect_to(moments_path) and return unless @bookmark.user_id == current_auth.user_id
+    
+    @bookmark.auth = current_auth
+    params[:bookmark][:color] = params[:highlight][:color] if params[:highlight]
     results = @bookmark.update(params[:bookmark])
     results.valid? ? redirect_to(:back, notice: t("bookmarks.successfully updated")) : render(action: "edit")
   end
 
   def destroy
     @bookmark = Bookmark.find(params[:id], auth: current_auth)
-    @bookmark.auth = current_auth
+    redirect_to(moments_path) and return unless @bookmark.user_id == current_auth.user_id
     
+    @bookmark.auth = current_auth
     results = @bookmark.destroy
     results.valid? ? redirect_to(user_bookmarks_path(current_user), notice: t("bookmarks.successfully deleted")) : render(action: "index")
   end
